refactor: Move color options to a shared location
